Moved provider parameters
